rousefleet/requirements.txt
google-cloud-aiplatform[adk,agent_engines]==1.91.0
Flask==3.1.0
python-dotenv==1.0.1
python-dateutil==2.8.2
Flask==3.1.0
google-cloud-spanner==3.54.0
humanize==4.12.3